requirements.txt
Pillow
wheel